fix: MVC 패턴으로 패키지 구조 변경  * 비지니스 로직을 가지는 객체를 domain 패키지로 이관  * UI 관련한 객체를 view 패키지로 이관
