+++ src/main/java/racingcar/domain/Cars.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.dto.CarResult;
 
 import java.util.ArrayList;
 import java.util.List;
