+++ src/main/java/racingcar/domain/Round.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Round {
