+++ src/main/java/racingcar/dto/GameResult.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 import java.util.ArrayList;
 import java.util.List;
