+++ src/main/java/racingcar/dto/RoundResult.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 import java.util.List;
