+++ src/main/java/racingcar/view/Billboard.java
@@ -1,4 +1,8 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.dto.CarResult;
+import racingcar.dto.GameResult;
+import racingcar.dto.RoundResult;
 
 import java.util.List;
 import java.util.stream.Collectors;
